rename config.py -> configuration.py
This should help cut down on the confusion between the "config" module
and the "config" variable that everyone likes to use. The config
variable should refer to an instance of Config, whereas before it was
being shared as both the name of the module and the name of an instance.

"configuration" is always the module. "config" should always be a Config
instance.

TODO: avoid passing around a Config instance everywhere.
