+++ pokemontools/__init__.py
@@ -1,3 +1,3 @@
-import config
+import configuration as config
 import crystal
 import preprocessor
